Fix mysql timestamp issues

+++ presto-flight-shim/src/main/java/com/facebook/presto/flightshim/ArrowBatchSource.java
@@ -608,7 +608,7 @@ public class ArrowBatchSource
         @Override
         public void writeLong(int index, long value)
         {
-            vector.set(index, (int) value);
+            vector.set(index, value);
         }
     }
 }
